fix(episode): unexpected episode type 255

+++ internal/mocks/EpisodeRepo.go
@@ -71,18 +71,18 @@ func (_c *EpisodeRepo_Count_Call) Return(_a0 int64, _a1 error) *EpisodeRepo_Coun
 }
 
 // CountByType provides a mock function with given fields: ctx, subjectID, epType
-func (_m *EpisodeRepo) CountByType(ctx context.Context, subjectID uint32, epType int8) (int64, error) {
+func (_m *EpisodeRepo) CountByType(ctx context.Context, subjectID uint32, epType int16) (int64, error) {
 	ret := _m.Called(ctx, subjectID, epType)
 
 	var r0 int64
-	if rf, ok := ret.Get(0).(func(context.Context, uint32, int8) int64); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, uint32, int16) int64); ok {
 		r0 = rf(ctx, subjectID, epType)
 	} else {
 		r0 = ret.Get(0).(int64)
 	}
 
 	var r1 error
-	if rf, ok := ret.Get(1).(func(context.Context, uint32, int8) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, uint32, int16) error); ok {
 		r1 = rf(ctx, subjectID, epType)
 	} else {
 		r1 = ret.Error(1)
@@ -99,14 +99,14 @@ type EpisodeRepo_CountByType_Call struct {
 // CountByType is a helper method to define mock.On call
 //  - ctx context.Context
 //  - subjectID uint32
-//  - epType int8
+//  - epType int16
 func (_e *EpisodeRepo_Expecter) CountByType(ctx interface{}, subjectID interface{}, epType interface{}) *EpisodeRepo_CountByType_Call {
 	return &EpisodeRepo_CountByType_Call{Call: _e.mock.On("CountByType", ctx, subjectID, epType)}
 }
 
-func (_c *EpisodeRepo_CountByType_Call) Run(run func(ctx context.Context, subjectID uint32, epType int8)) *EpisodeRepo_CountByType_Call {
+func (_c *EpisodeRepo_CountByType_Call) Run(run func(ctx context.Context, subjectID uint32, epType int16)) *EpisodeRepo_CountByType_Call {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(context.Context), args[1].(uint32), args[2].(int8))
+		run(args[0].(context.Context), args[1].(uint32), args[2].(int16))
 	})
 	return _c
 }
@@ -211,11 +211,11 @@ func (_c *EpisodeRepo_List_Call) Return(_a0 []model.Episode, _a1 error) *Episode
 }
 
 // ListByType provides a mock function with given fields: ctx, subjectID, epType, limit, offset
-func (_m *EpisodeRepo) ListByType(ctx context.Context, subjectID uint32, epType int8, limit int, offset int) ([]model.Episode, error) {
+func (_m *EpisodeRepo) ListByType(ctx context.Context, subjectID uint32, epType int16, limit int, offset int) ([]model.Episode, error) {
 	ret := _m.Called(ctx, subjectID, epType, limit, offset)
 
 	var r0 []model.Episode
-	if rf, ok := ret.Get(0).(func(context.Context, uint32, int8, int, int) []model.Episode); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, uint32, int16, int, int) []model.Episode); ok {
 		r0 = rf(ctx, subjectID, epType, limit, offset)
 	} else {
 		if ret.Get(0) != nil {
@@ -224,7 +224,7 @@ func (_m *EpisodeRepo) ListByType(ctx context.Context, subjectID uint32, epType 
 	}
 
 	var r1 error
-	if rf, ok := ret.Get(1).(func(context.Context, uint32, int8, int, int) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, uint32, int16, int, int) error); ok {
 		r1 = rf(ctx, subjectID, epType, limit, offset)
 	} else {
 		r1 = ret.Error(1)
@@ -241,16 +241,16 @@ type EpisodeRepo_ListByType_Call struct {
 // ListByType is a helper method to define mock.On call
 //  - ctx context.Context
 //  - subjectID uint32
-//  - epType int8
+//  - epType int16
 //  - limit int
 //  - offset int
 func (_e *EpisodeRepo_Expecter) ListByType(ctx interface{}, subjectID interface{}, epType interface{}, limit interface{}, offset interface{}) *EpisodeRepo_ListByType_Call {
 	return &EpisodeRepo_ListByType_Call{Call: _e.mock.On("ListByType", ctx, subjectID, epType, limit, offset)}
 }
 
-func (_c *EpisodeRepo_ListByType_Call) Run(run func(ctx context.Context, subjectID uint32, epType int8, limit int, offset int)) *EpisodeRepo_ListByType_Call {
+func (_c *EpisodeRepo_ListByType_Call) Run(run func(ctx context.Context, subjectID uint32, epType int16, limit int, offset int)) *EpisodeRepo_ListByType_Call {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(context.Context), args[1].(uint32), args[2].(int8), args[3].(int), args[4].(int))
+		run(args[0].(context.Context), args[1].(uint32), args[2].(int16), args[3].(int), args[4].(int))
 	})
 	return _c
 }
